streamlit-app/requirements.txt
altair==4.2.0
attrs==22.1.0
autopep8==1.7.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
DAWG-Python==0.7.2
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
faiss-cpu==1.7.2
filelock==3.8.0
flatbuffers==22.9.24
fonttools==4.37.4
gitdb==4.0.9
GitPython==3.1.27
graphviz==0.20.1
huggingface-hub==0.10.0
humanfriendly==10.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.9.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.0
mpmath==1.2.1
nltk==3.7
numpy==1.23.3
onnxruntime==1.12.1
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
plotly==5.10.0
protobuf==3.20.3
pyarrow==9.0.0
pybind11==2.10.0
pycodestyle==2.9.1
pydeck==0.8.0b3
Pygments==2.13.0
pyjarowinkler==1.8
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
Pympler==1.0.1
pyngrok==5.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
rich==12.5.1
scipy==1.9.1
semver==2.13.0
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
streamlit==1.13.0
streamlit-aggrid==0.3.3
sympy==1.11.1
tenacity==8.1.0
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1
tornado==6.2
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
zipp==3.8.1
